Animation displacement field across layer
For different values of the inhomogeneity factor, n, this GIF displays the evolution of the total displacement across the deposit for different values of the excitation base (normalized with respect to the fundamental frequency of a homogeneous site with constant base properties). This animation is equivalent to the second tile of GarciaSuarez2020.gif, which can also be found in this repository.